+++ NewPlatform.Flexberry.ORM.IntegratedTests/ICSSoft.STORMNET.Business/SQLDataServiceTest.Performance.cs
@@ -6,6 +6,7 @@
 
     using ICSSoft.STORMNET;
     using ICSSoft.STORMNET.Business;
+    using ICSSoft.STORMNET.FunctionalLanguage;
 
     using NewPlatform.Flexberry.ORM.Tests;
 
@@ -16,8 +17,96 @@
     /// </summary>
     public partial class SQLDataServiceTest
     {
+        private const int TestObjectCount = 1000;
+
         /// <summary>
-        /// Тест удаления детейлов.
+        /// Тест создания объектов без детейловых записей.
+        /// </summary>
+        [Fact(Skip = "Для ручного тестирования")]
+        public void CreateAgregatorWithoutDetailsTest()
+        {
+            foreach (IDataService dataService in DataServices)
+            {
+                // Arrange.
+                IEnumerable<int> range = Enumerable.Range(0, TestObjectCount);
+                var author = new Пользователь
+                {
+                    Логин = "BrainStorm",
+                };
+                dataService.UpdateObject(author);
+                var contest = new Конкурс
+                {
+                    Название = "Мозговой штурм",
+                    Организатор = author,
+                };
+                dataService.UpdateObject(contest);
+
+                // Act.
+                DataObject[] objs = range.Select(
+                    x => new Идея
+                    {
+                        СуммаБаллов = x,
+                        Автор = author,
+                        Конкурс = contest,
+                    }).ToArray();
+
+                Stopwatch stopwatch = new Stopwatch();
+                stopwatch.Start();
+                dataService.UpdateObjects(ref objs);
+                stopwatch.Stop();
+
+                output.WriteLine($"{nameof(CreateAgregatorWithoutDetailsTest)}@{dataService.GetType().Name}: elapsed {stopwatch.ElapsedMilliseconds}");
+            }
+        }
+
+        /// <summary>
+        /// Тест чтения объектов без детейловых записей.
+        /// </summary>
+        [Fact(Skip = "Для ручного тестирования")]
+        public void ReadAgregatorWithoutDetailsTest()
+        {
+            foreach (IDataService dataService in DataServices)
+            {
+                // Arrange.
+                IEnumerable<int> range = Enumerable.Range(0, TestObjectCount);
+                var author = new Пользователь
+                {
+                    Логин = "BrainStorm",
+                };
+                dataService.UpdateObject(author);
+                var contest = new Конкурс
+                {
+                    Название = "Мозговой штурм",
+                    Организатор = author,
+                };
+                dataService.UpdateObject(contest);
+
+                // Генерируем много объектов с детейлами.
+                DataObject[] objs = range.Select(
+                    x => new Идея
+                    {
+                        СуммаБаллов = x,
+                        Автор = author,
+                        Конкурс = contest,
+                    }).ToArray();
+                dataService.UpdateObjects(ref objs);
+
+                var lcs = LoadingCustomizationStruct.GetSimpleStruct(typeof(Идея), Идея.Views.ИдеяE);
+                lcs.LimitFunction = FunctionBuilder.BuildAnd(
+                    FunctionBuilder.BuildEquals<Идея>(x => x.Автор, author),
+                    FunctionBuilder.BuildEquals<Идея>(x => x.Конкурс, contest));
+
+                Stopwatch stopwatch = new Stopwatch();
+                stopwatch.Start();
+                var loadedObjs = dataService.LoadObjects(lcs);
+                stopwatch.Stop();
+
+                output.WriteLine($"{nameof(ReadAgregatorWithoutDetailsTest)}@{dataService.GetType().Name}: elapsed {stopwatch.ElapsedMilliseconds}");
+            }
+        }
+
+        /// <summary>
+        /// Тест удаления объектов с детейлами, но без детейловых записей.
         /// </summary>
         [Fact(Skip = "Для ручного тестирования")]
         public void DeleteNotLoadedAgregatorWithDetailsTest()
@@ -25,8 +114,7 @@
             foreach (IDataService dataService in DataServices)
             {
                 // Arrange.
-                const int length = 1000;
-                IEnumerable<int> range = Enumerable.Range(0, length);
+                IEnumerable<int> range = Enumerable.Range(0, TestObjectCount);
                 var author = new Пользователь
                 {
                     Логин = "BrainStorm",